requirements.txt
allure-pytest==2.8.18
click==7.1.2
jsonpath==0.82
openpyxl==3.0.5
pytest==6.0.1
pytest-html==3.0.0
requests==2.24.0
urllib3==1.25.7
zipp==3.1.0
cv2